Fix cmd line so it is type as it is seen to be
I tried using this line and it made a .env in the right place but it included the "" in the file. I feel like it help to not mislead people to just have the line typed out here as they would do it in their own cmd to local copy for better testing and contributing that you desire. If u copy and paste the line right now the way it is with the "" then u still get the error about double colon in host identifier. removing the "" fixes the issue.