Streamline Currently Required .env file creation
Coz: to double check, why would it be bad to automatically include a default value .env file in the master github code for cloning purposes to save installation confusion? Since this file is never appended onto after creation and everyone initializes theirs the exact same way as each other what would be leaked if this was already created and inserted into the main code in advance?