Timed games will stop when too many players leave
Specifically, election votes will not auto-pick if there are not enough players in the game to remake the game. Once this happens, all players need to manually vote in order to resume the game.

This was added to deal with games silently auto-completing despite being removed from the list of active games, as well as to combat players sitting in an otherwise abandoned game.