Reductions in data sent in the game list event
A bunch of true/false values are set to undefined when false to omit them (only send when true).
Timed mode is also omitted when 0, as 0 means no timer.
Passwords are omitted when empty, as that means no password.

A lot of games have tons of false values being sent, so this should be a notable improvement.