Code Cleanup (#1081)
* Code Cleanup

### Simplified convertToHex
See below for unit test.
```
let b = num => num.toString(16).padStart(4, "0")
for (let a = 0; a < 16 * 16 * 16 * 16; a++) {
if(convertToHex(a) !== b(a))console.error(`Failed for ${a}`);
}
```

### Simplified objIP
Rewrote the function with JS stream syntax

### Rewrote map generation
The old code was slow because it was constantly removing elements from the middle of arrays. This is inefficient because we have to shift the elements in front everytime we do this. It's a lot faster to first randomize the array and then process it sequentially. This takes around 200 ms on my Surface to execute.

* Added username to contributors