Fixed cardbacks silently failing over an arbitrary size that was below 100kb
Note that the 100kb check in index.js was removed, the length of that data is larger than the original image due to base64 encoding and therefore incorrectly failed images near the limit.

The limit is now enforced in app.js, set to 200kb to account for this encoding (100kb should become 125kb, 200kb is to be safe).

Additionally, a small limit was added to json-based requests. Realistically this should never be exceeded, as it should be on inbound messages only and should not apply to the socket.