Fix so that altmods and trialmods have colors
Currently in playerlist and chats these staffRoles (trialmod & altmod) do not stay earned elo colors and as a result default to grey color when assigned so anonymity of these roles is not retained. This patch will fix this error in the code that I forgot when switching these roles from hardcoded values into the database for assignment convenience.