improve language filter to prevent false positives
https://cdn.discordapp.com/attachments/326820032116162561/566908252403793930/htestresults.png
^
tests on strings with false positives, positives, and negatives for control.
added a few blacklisted words: f4g (previously false positive on "if 4 gets" or "if 4 got", so was removed), m0ng, t4rd.
In testing, speed shows approx. 1ms difference on false positive correction no matter the length of the string. This latency could likely be removed if:
A) exceptions becomes an Object like badWords, and
B) exceptions are tested either with, or before badWords. (e.g. if exception is found, that region of string does not test for bad words.

I've added notation all throughout my edits to clarify exactly what is occuring.

~Syr1nge